fix: handling request errors with non-existent ids

+++ controllers/cards.js
@@ -38,6 +38,12 @@ const deleteCard = async (req, res) => {
   // eslint-disable-next-line no-console
   console.log('deleteCard');
   const { cardId } = req.params;
+  if (cardId.length !== 24) {
+    const err = new Error('');
+    err.name = 'BadRequestError';
+    handleError(err, res);
+    return;
+  }
   try {
     const card = await Card.findByIdAndRemove(cardId);
     if (!card) {
@@ -55,6 +61,12 @@ const likeCard = async (req, res) => {
   // eslint-disable-next-line no-console
   console.log('likeCard');
   const { cardId } = req.params;
+  if (cardId.length !== 24) {
+    const err = new Error('');
+    err.name = 'BadRequestError';
+    handleError(err, res);
+    return;
+  }
   const ownerId = req.user._id;
   try {
     const card = await Card.findByIdAndUpdate(
@@ -62,6 +74,10 @@ const likeCard = async (req, res) => {
       { $addToSet: { likes: ownerId } }, // добавить _id в массив, если его там нет
       { new: true },
     );
+    if (!card) {
+      res.status(404).send({ message: 'Ошибка 404. Карточка не найдена' });
+      return;
+    }
     res.status(201).send(card);
   } catch (err) {
     handleError(err, res);
@@ -73,6 +89,12 @@ const deleteLike = async (req, res) => {
   // eslint-disable-next-line no-console
   console.log('deleteLike');
   const { cardId } = req.params;
+  if (cardId.length !== 24) {
+    const err = new Error('');
+    err.name = 'BadRequestError';
+    handleError(err, res);
+    return;
+  }
   const ownerId = req.user._id;
   try {
     const card = await Card.findByIdAndUpdate(
@@ -80,6 +102,10 @@ const deleteLike = async (req, res) => {
       { $pull: { likes: ownerId } }, // убрать _id из массива
       { new: true },
     );
+    if (!card) {
+      res.status(404).send({ message: 'Ошибка 404. Карточка не найдена' });
+      return;
+    }
     res.status(200).send(card);
   } catch (err) {
     handleError(err, res);
